# mec/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: ETSI GS MEC 013 - Location API
  description: The ETSI MEC ISG MEC013 Location API described using OpenAPI.
  contact:
    url: https://forge.etsi.org/rep/mec/gs013-location-api
  license:
    name: BSD-3-Clause
    url: https://forge.etsi.org/legal-matters
  version: 2.2.1
...
servers:
- url: https://localhost/mec_platform
...
paths:
  /queries/distance:
    get:
      tags:
      - location
      summary: UE Distance Lookup of a specific UE
      description: UE Distance Lookup between terminals or a terminal and a location
      operationId: distance_get
      parameters:
      - name: requester
        in: query
        description: Entity that is requesting the information
        required: false
        style: form
        explode: true
        schema:
          type: string
        x-exportParamName: Requester
      - name: address
        in: query
        description: "address of users (e.g. \"sip\" URI, \"tel\" URI, \"acr\" URI)"
        required: true
        style: form
        explode: true
        schema:
          type: array
          items:
            type: string
        x-exportParamName: AddressMandatory
      - name: latitude
        in: query
        description: Latitude geo position
        required: false
        style: form
        explode: true
        schema:
          type: number
          format: float
        x-exportParamName: Latitude
      - name: longitude
        in: query
        description: Longitude geo position
        required: false
        style: form
        explode: true
        schema:
          type: number
          format: float
        x-exportParamName: Longitude
      responses:
        "200":
          description: Successful response to a distance request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
        "400":
          description: "Bad Request : used to indicate that incorrect parameters were\
            \ passed to the request."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "401":
          description: "Unauthorized :  used when the client did not submit credentials."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "403":
          description: "Forbidden :  operation is not allowed given the current status\
            \ of the resource."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "404":
          description: "Not Found :  used when a client provided a URI that cannot\
            \ be mapped to a valid resource URI."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "406":
          description: "Not Acceptable : used to indicate that the server cannot provide\
            \ the any of the content formats supported by the client."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "429":
          description: "Too Many Requests : used when a rate limiter has triggered."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: queries
      x-openapi-router-controller: swagger_server.controllers.location_controller
  /queries/closestUser:
    get:
      tags:
      - location
      summary: Find UE closest to a location.
      description: Find the UE that is closest to another UE or a location.
      operationId: find_closest
      parameters:
      - name: address
        in: query
        description: "address of users (e.g. \"sip\" URI, \"tel\" URI, \"acr\" URI)"
        required: false
        style: form
        explode: true
        schema:
          type: string
        x-exportParamName: AddressMandatory
      - name: latitude
        in: query
        description: Latitude geo position
        required: false
        style: form
        explode: true
        schema:
          type: number
          format: float
        x-exportParamName: Latitude
      - name: longitude
        in: query
        description: Longitude geo position
        required: false
        style: form
        explode: true
        schema:
          type: number
          format: float
        x-exportParamName: Longitude
      - name: user_type
        in: query
        description: "Type of UE to search for"
        required: false
        style: form
        explode: true
        schema:
          type: array
          items:
            type: string
        x-exportParamName: UserType
      responses:
        "200":
          description: Successful response to closest user request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_1'
        "400":
          description: "Bad Request : used to indicate that incorrect parameters were\
            \ passed to the request."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "401":
          description: "Unauthorized :  used when the client did not submit credentials."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "403":
          description: "Forbidden :  operation is not allowed given the current status\
            \ of the resource."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "404":
          description: "Not Found :  used when a client provided a URI that cannot\
            \ be mapped to a valid resource URI."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "406":
          description: "Not Acceptable : used to indicate that the server cannot provide\
            \ the any of the content formats supported by the client."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "429":
          description: "Too Many Requests : used when a rate limiter has triggered."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: queries
      x-openapi-router-controller: swagger_server.controllers.location_controller
  /queries/closeUsers:
    get:
      tags:
      - location
      summary: Find UE closest to a location.
      description: Find the UE that is closest to another UE or a location.
      operationId: find_close
      parameters:
      - name: address
        in: query
        description: "address of users (e.g. \"sip\" URI, \"tel\" URI, \"acr\" URI)"
        required: false
        style: form
        explode: true
        schema:
          type: string
        x-exportParamName: AddressMandatory
      - name: latitude
        in: query
        description: Latitude geo position
        required: false
        style: form
        explode: true
        schema:
          type: number
          format: float
        x-exportParamName: Latitude
      - name: longitude
        in: query
        description: Longitude geo position
        required: false
        style: form
        explode: true
        schema:
          type: number
          format: float
        x-exportParamName: Longitude
      - name: distance
        in: query
        description: Maximum distance from target
        required: false
        style: form
        explode: true
        schema:
          type: number
          format: float
        x-exportParamName: Distance
      - name: user_type
        in: query
        description: "Type of UE to search for"
        required: false
        style: form
        explode: true
        schema:
          type: array
          items:
            type: string
        x-exportParamName: UserType
      responses:
        "200":
          description: Successful response to closest user request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_1'
        "400":
          description: "Bad Request : used to indicate that incorrect parameters were\
            \ passed to the request."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "401":
          description: "Unauthorized :  used when the client did not submit credentials."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "403":
          description: "Forbidden :  operation is not allowed given the current status\
            \ of the resource."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "404":
          description: "Not Found :  used when a client provided a URI that cannot\
            \ be mapped to a valid resource URI."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "406":
          description: "Not Acceptable : used to indicate that the server cannot provide\
            \ the any of the content formats supported by the client."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "429":
          description: "Too Many Requests : used when a rate limiter has triggered."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: queries
      x-openapi-router-controller: swagger_server.controllers.location_controller
  /queries/users:
    get:
      tags:
      - location
      summary: UE Location Lookup of a specific UE or group of UEs
      description: UE Location Lookup of a specific UE or group of UEs
      operationId: users_get
      parameters:
      - name: zoneId
        in: query
        description: Identifier of zone
        required: false
        style: form
        explode: true
        schema:
          type: array
          items:
            type: string
        x-exportParamName: ZoneId
      - name: accessPointId
        in: query
        description: Identifier of access point
        required: false
        style: form
        explode: true
        schema:
          type: array
          items:
            type: string
        x-exportParamName: AccessPointId
      - name: address
        in: query
        description: "address of users (e.g. \"sip\" URI, \"tel\" URI, \"acr\" URI)"
        required: false
        style: form
        explode: true
        schema:
          type: array
          items:
            type: string
        x-exportParamName: Address
      responses:
        "200":
          description: Successful response to users request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_1'
        "400":
          description: "Bad Request : used to indicate that incorrect parameters were\
            \ passed to the request."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "401":
          description: "Unauthorized :  used when the client did not submit credentials."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "403":
          description: "Forbidden :  operation is not allowed given the current status\
            \ of the resource."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "404":
          description: "Not Found :  used when a client provided a URI that cannot\
            \ be mapped to a valid resource URI."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "406":
          description: "Not Acceptable : used to indicate that the server cannot provide\
            \ the any of the content formats supported by the client."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "429":
          description: "Too Many Requests : used when a rate limiter has triggered."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: queries
      x-openapi-router-controller: swagger_server.controllers.location_controller
  /queries/zones:
    get:
      tags:
      - location
      summary: Zones information Lookup
      description: Used to get a list of identifiers for zones authorized for use
        by the application.
      operationId: zones_get
      responses:
        "200":
          description: Successful response to zones request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_2'
        "400":
          description: "Bad Request : used to indicate that incorrect parameters were\
            \ passed to the request."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "401":
          description: "Unauthorized :  used when the client did not submit credentials."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "403":
          description: "Forbidden :  operation is not allowed given the current status\
            \ of the resource."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "404":
          description: "Not Found :  used when a client provided a URI that cannot\
            \ be mapped to a valid resource URI."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "406":
          description: "Not Acceptable : used to indicate that the server cannot provide\
            \ the any of the content formats supported by the client."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "429":
          description: "Too Many Requests : used when a rate limiter has triggered."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: queries
      x-openapi-router-controller: swagger_server.controllers.location_controller
  /queries/zones/{zoneId}:
    get:
      tags:
      - location
      summary: Zones information Lookup
      description: Used to get the information for an authorized zone for use by the
        application.
      operationId: zones_get_by_id
      parameters:
      - name: zoneId
        in: path
        description: Indentifier of zone
        required: true
        style: simple
        explode: false
        schema:
          type: string
        x-exportParamName: ZoneId
      responses:
        "200":
          description: Successful response to an access point Id request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_3'
        "400":
          description: "Bad Request : used to indicate that incorrect parameters were\
            \ passed to the request."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "401":
          description: "Unauthorized :  used when the client did not submit credentials."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "403":
          description: "Forbidden :  operation is not allowed given the current status\
            \ of the resource."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "404":
          description: "Not Found :  used when a client provided a URI that cannot\
            \ be mapped to a valid resource URI."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "406":
          description: "Not Acceptable : used to indicate that the server cannot provide\
            \ the any of the content formats supported by the client."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "429":
          description: "Too Many Requests : used when a rate limiter has triggered."
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: queries
      x-openapi-router-controller: swagger_server.controllers.location_controller
...
components:
  schemas:
...
    TerminalDistance:
      required:
      - distance
      type: object
      properties:
        accuracy:
          type: integer
          description: Accuracy of the provided distance in meters
          x-etsi-mec-cardinality: 0..1
          x-etsi-mec-origin-type: integer
        distance:
          type: integer
          description: Distance from terminal to a location or between two terminals
            specified in meters
          x-etsi-mec-cardinality: 1
          x-etsi-mec-origin-type: integer
        timestamp:
          $ref: '#/components/schemas/TimeStamp'
      description: "A type containing information about the distance from a terminal\
        \ to a location or between two terminals, in addition the accuracy and a timestamp\
        \ of the information are provided."
      example:
        distance: 6
        accuracy: 0
        timestamp:
          seconds: 5
          nanoSeconds: 1
...
    inline_response_200:
      type: object
      properties:
        terminalDistance:
          $ref: '#/components/schemas/TerminalDistance'
      example:
        terminalDistance:
          distance: 6
          accuracy: 0
          timestamp:
            seconds: 5
            nanoSeconds: 1
    inline_response_200_1:
      type: object
      properties:
        userList:
          $ref: '#/components/schemas/UserList'
      example:
        userList:
          resourceURL: resourceURL
          user:
          - locationInfo:
              altitude: 5.962134
              orientationMajorAxis: 4
              shape: 7
              confidence: 5
              latitude:
              - 9.301444
              - 9.301444
              offsetAngle: 2
              accuracy: 0
              includedAngle: 2
              velocity:
                verticalUncertainty: 4
                horizontalSpeed: 1
                bearing: 1
                uncertainty: 6
                verticalSpeed: 1
                velocityType: 7
              accuracySemiMinor: 1
              innerRadius: 7
              accuracyAltitude: 6
              uncertaintyRadius: 1
              longitude:
              - 3.6160767
              - 3.6160767
              timestamp:
                seconds: 5
                nanoSeconds: 1
            address: http://example.com/aeiou
            resourceURL: resourceURL
            accessPointId: accessPointId
            zoneId: zoneId
            ancillaryInfo: ancillaryInfo
            contextLocationInfo: contextLocationInfo
            timestamp: null
          - locationInfo:
              altitude: 5.962134
              orientationMajorAxis: 4
              shape: 7
              confidence: 5
              latitude:
              - 9.301444
              - 9.301444
              offsetAngle: 2
              accuracy: 0
              includedAngle: 2
              velocity:
                verticalUncertainty: 4
                horizontalSpeed: 1
                bearing: 1
                uncertainty: 6
                verticalSpeed: 1
                velocityType: 7
              accuracySemiMinor: 1
              innerRadius: 7
              accuracyAltitude: 6
              uncertaintyRadius: 1
              longitude:
              - 3.6160767
              - 3.6160767
              timestamp:
                seconds: 5
                nanoSeconds: 1
            address: http://example.com/aeiou
            resourceURL: resourceURL
            accessPointId: accessPointId
            zoneId: zoneId
            ancillaryInfo: ancillaryInfo
            contextLocationInfo: contextLocationInfo
            timestamp: null
    inline_response_200_2:
      type: object
      properties:
        zoneList:
          $ref: '#/components/schemas/ZoneList'
      example:
        zoneList:
          resourceURL: resourceURL
          zone:
          - resourceURL: resourceURL
            numberOfUsers: 1
            zoneId: zoneId
            numberOfAccessPoints: 0
            numberOfUnserviceableAccessPoints: 6
          - resourceURL: resourceURL
            numberOfUsers: 1
            zoneId: zoneId
            numberOfAccessPoints: 0
            numberOfUnserviceableAccessPoints: 6
    inline_response_200_3:
      type: object
      properties:
        zoneInfo:
          $ref: '#/components/schemas/ZoneInfo'
      example:
        zoneInfo:
          resourceURL: resourceURL
          numberOfUsers: 1
          zoneId: zoneId
          numberOfAccessPoints: 0
          numberOfUnserviceableAccessPoints: 6
...
    ProblemDetails:
      title: ProblemDetails
      type: object
      properties:
        type:
          type: string
          description: A URI reference according to IETF RFC 3986 that identifies
            the problem type
        title:
          type: string
          description: "A short, human-readable summary of the problem type"
        status:
          type: integer
          description: The HTTP status code for this occurrence of the problem
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of
            the problem
        instance:
          type: string
          description: A URI reference that identifies the specific occurrence of
            the problem
...